# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 'API Recette'
  description: 'Cette API participative permet de partager des recettes entre utilisateurs.'
  version: '1'
servers:
  -
    url: 'http://localhost:8000'
    description: 'Mon API'
paths:
  /categories:
    get:
      tags:
        - Categorie
      summary: 'Collection of Categorie'
      description: 'Get a collection of Categorie'
      operationId: 'App\Controller\CategorieController::index'
      parameters:
        -
          name: 'orderBy[nom]'
          in: query
          description: 'Order by nom'
          required: false
          schema:
            type: array
            items:
              type: string
              enum:
                - asc
                - desc
      responses:
        '200':
          description: 'Get a collection of Categorie'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategorieReadRecette'
        '404':
          $ref: '#/components/responses/notFound'
    post:
      tags:
        - Categorie
      summary: 'Add an item Categorie'
      description: 'Create an item of Categorie'
      operationId: 'App\Controller\CategorieController::add'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                nom:
                  type: string
                recette:
                  type: array
                  items: { $ref: '#/components/schemas/Recette' }
              type: object
      responses:
        '201':
          description: 'Create an item of Categorie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categorie'
        '404':
          $ref: '#/components/responses/notFound'
        '401':
          $ref: '#/components/responses/unauthorized'
        '400':
          $ref: '#/components/responses/badRequest'
  '/categories/{id}':
    get:
      tags:
        - Categorie
      summary: 'Item of Categorie'
      description: 'Get an item of Categorie'
      operationId: 'App\Controller\CategorieController::show'
      parameters:
        -
          name: id
          in: path
          description: 'ID de la ressource'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Get an item of Categorie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categorie'
        '404':
          $ref: '#/components/responses/notFound'
        '401':
          $ref: '#/components/responses/unauthorized'
    put:
      tags:
        - Categorie
      summary: 'Update an item Categorie'
      description: 'Update an item of Categorie'
      operationId: 'App\Controller\CategorieController::update'
      parameters:
        -
          name: id
          in: path
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                nom:
                  type: string
                recettes:
                  type: array
                  items: { $ref: '#/components/schemas/Recette' }
              type: object
      responses:
        '200':
          description: 'Update an item of categorie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categorie'
        '404':
          $ref: '#/components/responses/notFound'
        '401':
          $ref: '#/components/responses/unauthorized'
    delete:
      tags:
        - Categorie
      summary: 'Delete an item Categorie'
      description: 'Delete an item of Categorie'
      operationId: 'App\Controller\CategorieController::delete'
      parameters:
        -
          name: id
          in: path
          required: true
      responses:
        '204':
          description: 'Success - Item is deleted'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: Success }
                type: object
        '404':
          $ref: '#/components/responses/notFound'
        '401':
          $ref: '#/components/responses/unauthorized'
...
components:
  schemas:
...
    Categorie:
      properties:
        id:
          type: integer
        nom:
          type: string
        recettes:
          type: array
          items:
            $ref: '#/components/schemas/Recette'
      type: object
    Recette:
      properties:
        id:
          type: integer
        tempsPreparation:
          type: integer
          default: '55'
        cout:
          type: integer
        nbPersonne:
          type: integer
        dateCreation:
          type: string
          format: date-time
        nom:
          type: string
        public:
          type: boolean
        user:
          $ref: '#/components/schemas/UserGetCollectionRecette'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Categorie'
      type: object
    CategorieReadRecette:
      properties:
        nbPersonne:
          type: integer
        cout:
          type: integer
        tempsPreparation:
          type: integer
        id:
          type: integer
        nom:
          type: string
        dateCreation:
          type: string
          format: date-time
        public:
          type: boolean
        user:
          $ref: '#/components/schemas/UserGetCollectionRecette'
      type: object
...
    UserGetCollectionRecette:
      properties:
        firstname:
          type: string
        lastname:
          type: string
      type: object
  responses:
    notFound:
      description: 'This resource not found'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Cette ressource n''existe pas'
            type: object
    unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Vous devez être authentifié'
            type: object
...
    badRequest:
      description: 'Bad request'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'bad request'
            type: object
...
  securitySchemes:
    basic:
      type: http
      scheme: basic